#29 split PlacingDefinition to its own Generator
Create a own Generator for PlacingDefinition in a separate js-file with all already defined methods to create a placing list + extract all testcases to a separate testsuite
