+++ webapp/src/graphical-editor/model/generator/ConnectionDefinitionGenerator.js
@@ -1,16 +1,9 @@
 
+import PlacingDefinitionGenerator from './connectionDefinitionGenerator/PlacingDefinitionGenerator'
+
 class ConnectionDefinitionGenerator { 
 
     constructor() {
-        this.placingShape = {
-            'line': (shape) => this.generateLineShape(shape),
-            'polyline': (shape) => this.generatePolyLineShape(shape),
-            'polygon': (shape) => this.generatePolygonShape(shape),
-            'rectangle': (shape, distance) => this.generateRectangleShape(shape, distance),
-            'roundedRectangle': (shape, distance) => this.generateRoundedRectangleShape(shape, distance),
-            'ellipse': (shape, distance) => this.generateEllipseShape(shape, distance),
-            'text': (shape) => this.generateTextShape(shape)
-        };
 
         this.generateSvgPathData = {
             'line': (shape) => this.generateLineSvgPathData(shape),
@@ -124,7 +117,7 @@ class ConnectionDefinitionGenerator {
         const head = mirroredPoints[0];
         const tail = mirroredPoints.slice(1);
 
-        return ("M " + head.x + " " + head.y + " " + tail.map(p => "L " + p.x + " " + p.y)).replace(",", "");
+        return ("M " + head.x + " " + head.y + " " + tail.map(p => "L " + p.x + " " + p.y) + "z").replace(",", "");
     }
 
     generateLineSvgPathData(shape) {
@@ -174,108 +167,6 @@ class ConnectionDefinitionGenerator {
             }
         }
     }
-
-    createPlacingList(connection) {
-        return connection.placings.map(this.createPlacing, this);
-    }
-
-    createPlacing(placing) {
-        const generatedPlacing = {
-            position: placing.positionOffset
-        };
-        return Object.assign(generatedPlacing, this.createPlacingShape(placing));
-    }
-
-    createPlacingShape(placing) {
-
-        let placingShape = this.placingShape[placing.shape.type](placing.shape, placing.positionDistance);
-        placingShape.attrs = placing.shape.type !== 'text' && 'style' in placing.shape ? Object.assign(placingShape.attrs, {style: placing.shape.style}): placingShape.attrs;
-    
-        return placingShape;
-    }
-
-    generateLineShape(line) {
-        return {
-            markup: '<line />',
-            attrs: {
-                x1: line.startPoint.x,
-                y1: line.startPoint.y,
-                x2: line.endPoint.x,
-                y2: line.endPoint.y
-            }
-        };
-    }
-
-    generatePolyLineShape(shape) {
-        return {
-            markup: '<polyline />',
-            attrs: {
-                points: this.generatePoints(shape.points),
-                fill: 'transparent'
-            }
-        };          
-    }
-
-    generatePoints(points) {
-        let pointString = "";
-        points.map(function(point) {
-            pointString += (`${point.x},${point.y} `)
-        })
-        return pointString.trim();
-    }
-    
-    generateRectangleShape(rectangle, distance) {
-        return {
-            markup: '<rect />',
-            attrs:{
-                height: rectangle.sizeHeight,
-                width: rectangle.sizeWidth,
-                y: distance - rectangle.sizeHeight / 2
-            }
-        };        
-    }
-    
-    generateRoundedRectangleShape(roundedRectangle, distance) {
-        return {
-            markup: '<rect />',
-            attrs:{
-                height: roundedRectangle.sizeHeight,
-                width: roundedRectangle.sizeWidth,
-                rx: roundedRectangle.curveWidth,
-                ry: roundedRectangle.curveHeight,
-                y: distance - roundedRectangle.sizeHeight / 2,
-            }
-        }
-    }
-      
-    generatePolygonShape(polygon) {
-        return {
-            markup: '<polygon />',
-            attrs:{
-              points: this.generatePoints(polygon.points)          
-            }
-        }
-    }
-    
-    generateEllipseShape(ellipse, distance) {
-        return {
-            markup: '<ellipse />',
-            attrs:{
-                rx: ellipse.sizeWidth / 2,
-                ry: ellipse.sizeHeight / 2,
-                cy: distance,
-            }
-        };
-    }
-    
-    generateTextShape(text) {
-        return {
-            markup: `<text>${text.textBody}</text>`,
-            attrs:{
-                y: text.sizeHeight / 2
-            }
-        };
-    }
     
     createLabelList(connection) {
         const labels = connection.placings.filter(placing => placing.shape.type === 'Label');
@@ -302,6 +193,7 @@ export default class Generator{
     constructor(connections) {
         // Braucht eine Uebergabe eines StyleGenerators
         this.connectionDefinitionGenerator = new ConnectionDefinitionGenerator();
+        this.placingDefinitionGenerator = new PlacingDefinitionGenerator(connections);
         this.connections = connections;
     }
 
@@ -312,7 +204,7 @@ export default class Generator{
 
     getPlacings(styleName) {
         const connection = this.connections.find(c => c.name === styleName);
-        return connection ? this.connectionDefinitionGenerator.createPlacingList(connection) : [];
+        return connection ? this.placingDefinitionGenerator.createPlacingList(connection) : [];
     }
 
     getLabels(styleName) {
